requirements.txt
# Core scientific stack
numpy>=1.26,<2.0
scipy>=1.11
pandas>=2.1

# Machine learning
scikit-learn>=1.3,<1.6

# NLP embeddings
sentence-transformers>=2.6.0
transformers>=4.40.0
torch>=2.2,<2.4

# Clustering utilities
kneed>=0.8.5

# Visualization
matplotlib>=3.8
seaborn>=0.13

# OpenAI API
openai>=1.30.0

# Environment variable management
python-dotenv>=1.0

# Notebook support
jupyter>=1.0
ipython>=8.0

# Progress bars
tqdm>=4.66

# Optional but recommended for clean CSV handling
openpyxl>=3.1